moved formSchema to formData file, changed application buttons to fit the text properly, form is not submitting!
