+++ app/applications/artisan/FormSections/AdditionalQuestions.tsx
@@ -5,8 +5,7 @@ import { Textarea } from "@/components/ui/textarea"
 import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group"
 import { UseFormReturn } from "react-hook-form";
 import { z } from "zod";
-
-import { formSchema } from '../page';
+import { formSchema } from '../formData';
 
 type FormValues = z.infer<typeof formSchema>;
