+++ app/applications/artisan/FormSections/GeneralInformation.tsx
@@ -5,8 +5,7 @@ import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@
 import { UseFormReturn } from "react-hook-form";
 import { z } from 'zod';
 
-import { formSchema } from '../page';
-import { states } from '../formData';
+import { states, formSchema } from '../formData';
 
 // Define the shape of your form values
 type FormValues = z.infer<typeof formSchema>;
