+++ app/applications/artisan/FormSections/PayrollInformation.tsx
@@ -5,8 +5,7 @@ import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@
 import { UseFormReturn } from "react-hook-form";
 import { z } from "zod";
 
-import { formSchema } from '../page';
-import { jobClasses } from '../formData';
+import { jobClasses, formSchema } from '../formData';
 
 type FormValues = z.infer<typeof formSchema>;
